# docs/openapi-public.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/orders": {
      "post": {
        "operationId": "OrdersController_createOrder",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_listOrders",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "paymentStatus",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fulfillmentStatus",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/convert-from-quote": {
      "post": {
        "operationId": "OrdersController_convertQuoteToOrder",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/stats": {
      "get": {
        "operationId": "OrdersController_getOrderStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}": {
      "get": {
        "operationId": "OrdersController_getOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      },
      "patch": {
        "operationId": "OrdersController_updateOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      },
      "delete": {
        "operationId": "OrdersController_deleteOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/line-items": {
      "post": {
        "operationId": "OrdersController_addLineItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/line-items/{lineItemId}": {
      "patch": {
        "operationId": "OrdersController_updateLineItem",
        "parameters": [
          {
            "name": "lineItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      },
      "delete": {
        "operationId": "OrdersController_removeLineItem",
        "parameters": [
          {
            "name": "lineItemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/fulfillment": {
      "patch": {
        "operationId": "OrdersController_updateFulfillment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/payment": {
      "patch": {
        "operationId": "OrdersController_updatePayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/cancel": {
      "post": {
        "operationId": "OrdersController_cancelOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/timeline": {
      "get": {
        "operationId": "OrdersController_getTimeline",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/confirm": {
      "post": {
        "operationId": "OrdersController_confirmOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/ship": {
      "post": {
        "operationId": "OrdersController_shipOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/orders/{id}/clone": {
      "post": {
        "operationId": "OrdersController_cloneOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/recalculate": {
      "post": {
        "operationId": "OrdersController_recalculateOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/pdf": {
      "get": {
        "operationId": "OrdersController_generatePdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/account/{accountId}": {
      "get": {
        "operationId": "OrdersController_getByAccountId",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "SalesOS API",
    "description": "SalesOS is an AI-powered Sales CRM and Revenue Intelligence platform. This API provides programmatic access to CRM data, sales coaching, CPQ, and more.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://www.salesos.org/api",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key authentication"
      }
    },
...
  }
}